+++ module/config.js
@@ -140,107 +140,6 @@ HERO.hitLocations = {
     Foot: [1, 0.5, 0.5, -8],
 };
 
-HERO.combatManeuvers = {
-    // Maneuver : [phase, OCV, DCV, Effects, Attack]
-    Block: ["1/2", "+0", "+0", "Blocks HTH attacks, Abort", true],
-    Brace: ["0", "+2", "1/2", "+2 OCV only to offset the Range Modifier"],
-    Disarm: [
-        "1/2",
-        "-2",
-        "+0",
-        "Disarm target, requires STR vs. STR Roll",
-        true,
-    ],
-    Dodge: ["1/2", "--", "+3", "Dodge all attacks, Abort", true],
-    Grab: [
-        "1/2",
-        "-1",
-        "-2",
-        "Grab Two Limbs; can Squeeze, Slam, or Throw",
-        true,
-    ],
-    "Grab By": [
-        "1/2 †",
-        "-3",
-        "-4",
-        "Move and Grab object, +(v/10) to STR",
-        true,
-    ],
-    Haymaker: ["1/2*", "+0", "-5", "+4 Damage Classes to any attack"],
-    "Move By": [
-        "1/2 †",
-        "-2",
-        "-2",
-        "((STR/2) + (v/10))d6; attacker takes 1/3 damage",
-        true,
-    ],
-    "Move Through": [
-        "1/2 †",
-        "-v/10",
-        "-3",
-        "(STR + (v/6))d6; attacker takes 1/2 or full damage",
-        true,
-    ],
-    //"Multiple Attack": ["1", "var", "1/2", "Attack one or more targets multiple times"],
-    Set: [
-        "1",
-        "+1",
-        "+0",
-        "Take extra time to aim a Ranged attack at a target",
-    ],
-    Shove: ["1/2", "-1", "-1", "Push target back 1m per 5 STR used", true],
-    Strike: ["1/2", "+0", "+0", "STR damage or by weapon type", true],
-    Throw: [
-        "1/2",
-        "+0",
-        "+0",
-        "Throw object or character, does STR damage",
-        true,
-    ],
-    Trip: [
-        "1/2",
-        "-1",
-        "-2",
-        "Knock a target to the ground, making him Prone",
-        true,
-    ],
-    //"Other Attacks": ["1/2", "+0", "+0", ""],
-};
-
-HERO.combatManeuversOptional = {
-    // Maneuver : [phase, OCV, DCV, Effects]
-    Choke: ["1/2", "-2", "-2", "NND 1d6, Grab One Limb"],
-    "Club Weapon": [
-        "1/2",
-        "+0",
-        "+0",
-        "Killing weapon does equivalent Normal Damage",
-    ],
-    Cover: ["1/2", "-2", "+0", "Target held at gunpoint"],
-    "Dive For Cover": ["1/2", "+0", "+0", "Character avoids attack; Abort"],
-    Hipshot: ["1/2", "-1", "+0", "+1", "DEX only for purposes of initiative"],
-    "Pulling A Punch": [
-        "1/2",
-        "-1/5d6",
-        "+0",
-        "Strike, normal STUN damage, 1/2 BODY damage",
-    ],
-    "Roll With A Punch": [
-        "1/2",
-        "-2",
-        "-2",
-        "Block after being hit, take 1/2 damage; Abort",
-    ],
-    "Snap Shot": ["1", "-1", "+0", "Lets character duck back behind cover"],
-    Strafe: ["1/2 †", "-v/6", "-2", "Make Ranged attack while moving"],
-    "Suppression Fire": [
-        "1/2",
-        "-2",
-        "+0",
-        "Continuous fire through an area, must be Autofire",
-    ],
-};
-
 // TODO: This could be created from powers.
 HERO.movementPowers = {
     extradimensionalmovement: "Extra Dimensional Movement",
@@ -270,7 +169,12 @@ HERO.powers5e = [];
  * @param {string} cost - Cost in character points per additional level
  * @param {Array<string>} type - A list of types associated with this power
  * @param {Array<"success">} behaviors - A list of the behavior types this power exhibits in the code
+ *                                       "non-hd" - this is not an XMLID that comes from Hero Designer
+ *                                       "optional-maneuver" - this is an optional combat maneuver
  *                                       "success" - can roll some kind of success roll for this power
+ *                                       "dice" - a damage/effect dice roll is associated with this power
+ *                                       "attack" - a to-hit dice roll is associated with this power
+ *                                       "on-off" - this power can be turned on or off
  *
  * @param {"constant"|"instant"|"persistent"} duration - The lower case duration of the power
  * @param {boolean} [costEnd] - If the power costs endurance to use. true if it does, false or undefined if it doesn't
@@ -388,7 +292,7 @@ function addPower(powerDescription6e, powerOverrideFor5e) {
         type: ["characteristic"],
         behaviors: ["success"],
         duration: "persistent",
-        ignoreFor: ["vehicle", "base2", "computer", "ai", "6e"],
+        ignoreFor: ["vehicle", "base2", "computer", "ai", "6e"], // TODO: Remove the 6e here.
         base: 10,
         cost: 1 / 2,
     });
@@ -696,6 +600,695 @@ function addPower(powerDescription6e, powerOverrideFor5e) {
             duration: "persistent",
         },
         {},
+    );
+})();
+
+(function addManeuversToPowerList() {
+    addPower(undefined, {
+        key: "BLAZINGAWAY",
+        type: ["maneuver"],
+        behaviors: ["non-hd", "optional-maneuver"],
+        name: "Blazing Away",
+        perceivability: "obvious",
+        duration: "instant",
+        costEnd: false,
+        target: "target's dcv",
+        ignoreFor: ["base2", "computer", "ai"],
+        maneuverDesc: {
+            phase: "1/2",
+            ocv: "+0",
+            dcv: "+0",
+            effects: "Make as many attacks as desired, only hit on a 3",
+            attack: true, // TODO: Do we want this property? Should likely be part of the behaviors. Same comment for all maneuvers.
+        },
+    });
+    addPower(
+        {
+            key: "BLOCK",
+            type: ["maneuver"],
+            behaviors: ["non-hd"],
+            name: "Block",
+            perceivability: "obvious",
+            duration: "instant",
+            costEnd: false,
+            target: "target's dcv",
+            ignoreFor: ["base2", "computer", "ai"],
+            maneuverDesc: {
+                phase: "1/2",
+                ocv: "+0",
+                dcv: "+0",
+                effects: "Blocks HTH attacks, Abort",
+                attack: true, // TODO: Should be false as it's not an attack. It does, however, require dice.
+            },
+        },
+        {},
+    );
+    addPower(
+        {
+            key: "BRACE",
+            type: ["maneuver"],
+            behaviors: ["non-hd"],
+            name: "Brace",
+            perceivability: "obvious",
+            duration: "instant",
+            costEnd: false,
+            target: "target's dcv",
+            ignoreFor: ["base2", "computer", "ai"],
+            maneuverDesc: {
+                phase: "0",
+                ocv: "+2",
+                dcv: "1/2",
+                effects: "+2 OCV only to offset the Range Modifier",
+                attack: false,
+            },
+        },
+        {},
+    );
+
+    addPower(
+        {
+            key: "CHOKE",
+            type: ["maneuver"],
+            behaviors: ["non-hd", "optional-maneuver"],
+            name: "Choke",
+            perceivability: "obvious",
+            duration: "instant",
+            costEnd: false,
+            target: "target's dcv",
+            ignoreFor: ["base2", "computer", "ai"],
+            maneuverDesc: {
+                phase: "1/2",
+                ocv: "-2",
+                dcv: "-2",
+                effects: "NND 1d6, Grab One Limb",
+                attack: true,
+            },
+        },
+        undefined,
+    );
+    addPower(
+        {
+            key: "CLUBWEAPON",
+            type: ["maneuver"],
+            behaviors: ["non-hd", "optional-maneuver"],
+            name: "Club Weapon",
+            perceivability: "obvious",
+            duration: "instant",
+            costEnd: false,
+            target: "target's dcv",
+            ignoreFor: ["base2", "computer", "ai"],
+            maneuverDesc: {
+                phase: "1/2",
+                ocv: "+0",
+                dcv: "+0",
+                effects: "Killing weapon does equivalent Normal Damage",
+                attack: true,
+            },
+        },
+        {},
+    );
+    addPower(
+        {
+            key: "COVER",
+            type: ["maneuver"],
+            behaviors: ["non-hd", "optional-maneuver"],
+            name: "Cover",
+            perceivability: "obvious",
+            duration: "instant",
+            costEnd: false,
+            target: "target's dcv",
+            ignoreFor: ["base2", "computer", "ai"],
+            maneuverDesc: {
+                phase: "1/2",
+                ocv: "-2",
+                dcv: "+0",
+                effects: "Target held at gunpoint",
+                attack: true,
+            },
+        },
+        {},
+    );
+
+    addPower(
+        {
+            key: "DISARM",
+            type: ["maneuver"],
+            behaviors: ["non-hd"],
+            name: "Disarm",
+            perceivability: "obvious",
+            duration: "instant",
+            costEnd: false,
+            target: "target's dcv",
+            ignoreFor: ["base2", "computer", "ai"],
+            maneuverDesc: {
+                phase: "1/2",
+                ocv: "-2",
+                dcv: "+0",
+                effects: "Disarm target, requires STR vs. STR Roll",
+                attack: true,
+            },
+        },
+        {},
+    );
+    addPower(
+        {
+            key: "DIVEFORCOVER",
+            type: ["maneuver"],
+            behaviors: ["non-hd", "optional-maneuver"],
+            name: "Dive For Cover",
+            perceivability: "obvious",
+            duration: "instant",
+            costEnd: false,
+            target: "target's dcv",
+            ignoreFor: ["base2", "computer", "ai"],
+            maneuverDesc: {
+                phase: "1/2",
+                ocv: "+0",
+                dcv: "+0",
+                effects: "Character avoids attack; Abort",
+                attack: true,
+            },
+        },
+        {},
+    );
+    addPower(
+        {
+            key: "DODGE",
+            type: ["maneuver"],
+            behaviors: ["non-hd"],
+            name: "Dodge",
+            perceivability: "obvious",
+            duration: "instant",
+            costEnd: false,
+            target: "target's dcv",
+            ignoreFor: ["base2", "computer", "ai"],
+            maneuverDesc: {
+                phase: "1/2",
+                ocv: "--",
+                dcv: "+3",
+                effects: "Dodge all attacks, Abort",
+                attack: true,
+            },
+        },
+        {},
+    );
+
+    addPower(
+        {
+            key: "GRAB",
+            type: ["maneuver"],
+            behaviors: ["non-hd"],
+            name: "Grab",
+            perceivability: "obvious",
+            duration: "instant",
+            costEnd: false,
+            target: "target's dcv",
+            ignoreFor: ["base2", "computer", "ai"],
+            maneuverDesc: {
+                phase: "1/2",
+                ocv: "-1",
+                dcv: "-2",
+                effects: "Grab Two Limbs; can Squeeze, Slam, or Throw",
+                attack: true,
+            },
+        },
+        {
+            maneuverDesc: {
+                phase: "1/2",
+                ocv: "-1",
+                dcv: "-2",
+                effects: "Grab Two Limbs; can Squeeze or Throw",
+                attack: true,
+            },
+        },
+    );
+    addPower(
+        {
+            key: "GRABBY",
+            type: ["maneuver"],
+            behaviors: ["non-hd"],
+            name: "Grab By",
+            perceivability: "obvious",
+            duration: "instant",
+            costEnd: false,
+            target: "target's dcv",
+            ignoreFor: ["base2", "computer", "ai"],
+            maneuverDesc: {
+                phase: "1/2 †",
+                ocv: "-3",
+                dcv: "-4",
+                effects: "Move and Grab object, +(v/10) to STR",
+                attack: true,
+            },
+        },
+        {
+            maneuverDesc: {
+                phase: "1/2 †",
+                ocv: "-3",
+                dcv: "-4",
+                effects: "Move and Grab object, +(v/5) to STR",
+                attack: true,
+            },
+        },
+    );
+
+    addPower(
+        {
+            key: "HAYMAKER",
+            type: ["maneuver"],
+            behaviors: ["non-hd"],
+            name: "Haymaker",
+            ignoreFor: ["base2", "computer", "ai"],
+            maneuverDesc: {
+                phase: "1/2 *",
+                ocv: "+0",
+                dcv: "-5",
+                effects: "+4 Damage Classes to any attack",
+                attack: false,
+            },
+        },
+        {},
+    );
+    addPower(
+        {
+            key: "HIPSHOT",
+            type: ["maneuver"],
+            behaviors: ["non-hd", "optional-maneuver"],
+            name: "Hipshot",
+            perceivability: "obvious",
+            duration: "instant",
+            costEnd: false,
+            target: "target's dcv",
+            ignoreFor: ["base2", "computer", "ai"],
+            maneuverDesc: {
+                phase: "1/2",
+                ocv: "-1",
+                dcv: "+0",
+                effects: "+1 DEX only for purposes of initiative",
+                attack: true,
+            },
+        },
+        {},
+    );
+    addPower(undefined, {
+        key: "HURRY",
+        type: ["maneuver"],
+        behaviors: ["non-hd", "optional-maneuver"],
+        name: "Hurry",
+        perceivability: "obvious",
+        duration: "instant",
+        costEnd: false,
+        target: "target's dcv",
+        ignoreFor: ["base2", "computer", "ai"],
+        maneuverDesc: {
+            phase: "1/2",
+            ocv: "-2",
+            dcv: "-2",
+            effects: "+1d6 DEX only for purposes of initiative",
+            attack: false,
+        },
+    });
+
+    addPower(
+        {
+            key: "MOVEBY",
+            type: ["maneuver"],
+            behaviors: ["non-hd"],
+            name: "Move By",
+            perceivability: "obvious",
+            duration: "instant",
+            costEnd: false,
+            target: "target's dcv",
+            ignoreFor: ["base2", "computer", "ai"],
+            maneuverDesc: {
+                phase: "1/2 †",
+                ocv: "-2",
+                dcv: "-2",
+                effects: "((STR/2) + (v/10))d6; attacker takes ⅓ damage",
+                attack: true,
+            },
+        },
+        {
+            maneuverDesc: {
+                phase: "1/2 †",
+                ocv: "-2",
+                dcv: "-2",
+                effects: "((STR/2) + (v/5))d6; attacker takes ⅓ damage",
+                attack: true,
+            },
+        },
+    );
+    addPower(
+        {
+            key: "MOVETHROUGH",
+            type: ["maneuver"],
+            behaviors: ["non-hd"],
+            name: "Move Through",
+            perceivability: "obvious",
+            duration: "instant",
+            costEnd: false,
+            target: "target's dcv",
+            ignoreFor: ["base2", "computer", "ai"],
+            maneuverDesc: {
+                phase: "1/2 †",
+                ocv: "-v/10",
+                dcv: "-3",
+                effects: "(STR + (v/6))d6; attacker takes ½ or full damage",
+                attack: true,
+            },
+        },
+        {
+            maneuverDesc: {
+                phase: "1/2 †",
+                ocv: "-v/5",
+                dcv: "-3",
+                effects: "(STR + (v/3))d6; attacker takes ½ or full damage",
+                attack: true,
+            },
+        },
+    );
+    addPower(
+        {
+            key: "MULTIPLEATTACK",
+            type: ["maneuver"],
+            behaviors: ["non-hd"],
+            name: "Multiple Attack",
+            perceivability: "obvious",
+            duration: "instant",
+            costEnd: false,
+            target: "target's dcv",
+            ignoreFor: ["base2", "computer", "ai"],
+            maneuverDesc: {
+                phase: "1",
+                ocv: "var",
+                dcv: "1/2",
+                effects: "Attack one or more targets multiple times",
+                attack: true,
+            },
+        },
+        undefined,
+    );
+
+    addPower(
+        {
+            key: "OTHERATTACKS",
+            type: ["maneuver"],
+            behaviors: ["non-hd"],
+            name: "Other Attacks",
+            perceivability: "obvious",
+            duration: "instant",
+            costEnd: false,
+            target: "target's dcv",
+            ignoreFor: ["base2", "computer", "ai"],
+            maneuverDesc: {
+                phase: "1/2",
+                ocv: "+0",
+                dcv: "+0",
+                effects: "For unlisted attacks",
+                attack: true,
+            },
+        },
+        undefined,
+    );
+
+    addPower(
+        {
+            key: "PULLINGAPUNCH",
+            type: ["maneuver"],
+            behaviors: ["non-hd", "optional-maneuver"],
+            name: "Pulling A Punch",
+            perceivability: "obvious",
+            duration: "instant",
+            costEnd: false,
+            target: "target's dcv",
+            ignoreFor: ["base2", "computer", "ai"],
+            maneuverDesc: {
+                phase: "1/2",
+                ocv: "-1/5d6",
+                dcv: "+0",
+                effects: "Strike, normal STUN damage, ½ BODY damage",
+                attack: true,
+            },
+        },
+        {},
+    );
+
+    addPower(undefined, {
+        key: "RAPIDFIRE",
+        type: ["maneuver"],
+        behaviors: ["non-hd", "optional-maneuver"],
+        name: "Rapid Fire",
+        perceivability: "obvious",
+        duration: "instant",
+        costEnd: false,
+        target: "target's dcv",
+        ignoreFor: ["base2", "computer", "ai"],
+        maneuverDesc: {
+            phase: "1",
+            ocv: "-1/x",
+            dcv: "+x1/2",
+            effects: "Strike, normal STUN damage, ½ BODY damage",
+            attack: true,
+        },
+    });
+    addPower(
+        {
+            key: "ROLLWITHAPUNCH",
+            type: ["maneuver"],
+            behaviors: ["non-hd", "optional-maneuver"],
+            name: "Roll With A Punch",
+            perceivability: "obvious",
+            duration: "instant",
+            costEnd: false,
+            target: "target's dcv",
+            ignoreFor: ["base2", "computer", "ai"],
+            maneuverDesc: {
+                phase: "1/2",
+                ocv: "-2",
+                dcv: "-2",
+                effects: "Block after being hit, take ½ damage; Abort",
+                attack: true,
+            },
+        },
+        {},
+    );
+
+    addPower(
+        {
+            key: "SET",
+            type: ["maneuver"],
+            behaviors: ["non-hd"],
+            name: "Set",
+            perceivability: "obvious",
+            duration: "instant",
+            costEnd: false,
+            target: "target's dcv",
+            ignoreFor: ["base2", "computer", "ai"],
+            maneuverDesc: {
+                phase: "1",
+                ocv: "+1",
+                dcv: "+0",
+                effects: "Take extra time to aim a Ranged attack at a target",
+                attack: false,
+            },
+        },
+        {},
+    );
+    addPower(
+        {
+            key: "SETANDBRACE",
+            type: ["maneuver"],
+            behaviors: ["non-hd"],
+            name: "Set And Brace",
+            perceivability: "obvious",
+            duration: "instant",
+            costEnd: false,
+            target: "target's dcv",
+            ignoreFor: ["base2", "computer", "ai"],
+            maneuverDesc: {
+                phase: "1",
+                ocv: "+3",
+                dcv: "1/2",
+                effects:
+                    "Take extra time to aim a Ranged attack at a target, +2 OCV only to offset the Range Modifier",
+                attack: false,
+            },
+        },
+        {},
+    );
+    addPower(
+        {
+            key: "SHOVE",
+            type: ["maneuver"],
+            behaviors: ["non-hd"],
+            name: "Shove",
+            perceivability: "obvious",
+            duration: "instant",
+            costEnd: false,
+            target: "target's dcv",
+            ignoreFor: ["base2", "computer", "ai"],
+            maneuverDesc: {
+                phase: "1/2",
+                ocv: "-1",
+                dcv: "-1",
+                effects: "Push target back 1m per 5 STR used",
+                attack: true,
+            },
+        },
+        undefined,
+    );
+    addPower(
+        {
+            key: "SNAPSHOT",
+            type: ["maneuver"],
+            behaviors: ["non-hd", "optional-maneuver"],
+            name: "Snap Shot",
+            perceivability: "obvious",
+            duration: "instant",
+            costEnd: false,
+            target: "target's dcv",
+            ignoreFor: ["base2", "computer", "ai"],
+            maneuverDesc: {
+                phase: "1",
+                ocv: "-1",
+                dcv: "+0",
+                effects: "Lets character duck back behind cover",
+                attack: true,
+            },
+        },
+        {},
+    );
+    addPower(
+        {
+            key: "STRAFE",
+            type: ["maneuver"],
+            behaviors: ["non-hd", "optional-maneuver"],
+            name: "Strafe",
+            perceivability: "obvious",
+            duration: "instant",
+            costEnd: false,
+            target: "target's dcv",
+            ignoreFor: ["base2", "computer", "ai"],
+            maneuverDesc: {
+                phase: "1/2 †",
+                ocv: "-v/6",
+                dcv: "-2",
+                effects: "Make Ranged attack while moving",
+                attack: true,
+            },
+        },
+        undefined,
+    );
+    addPower(
+        {
+            key: "STRIKE",
+            type: ["maneuver"],
+            behaviors: ["non-hd"],
+            name: "Strike",
+            perceivability: "obvious",
+            duration: "instant",
+            costEnd: false,
+            target: "target's dcv",
+            ignoreFor: ["base2", "computer", "ai"],
+            maneuverDesc: {
+                phase: "1/2",
+                ocv: "+0",
+                dcv: "+0",
+                effects: "STR damage or by weapon type",
+                attack: true,
+            },
+        },
+        {},
+    );
+    addPower(
+        {
+            key: "SUPPRESSIONFIRE",
+            type: ["maneuver"],
+            behaviors: ["non-hd", "optional-maneuver"],
+            name: "Suppression Fire",
+            perceivability: "obvious",
+            duration: "instant",
+            costEnd: false,
+            target: "target's dcv",
+            ignoreFor: ["base2", "computer", "ai"],
+            maneuverDesc: {
+                phase: "1/2",
+                ocv: "-2",
+                dcv: "+0",
+                effects: "Continuous fire through an area, must be Autofire",
+                attack: true,
+            },
+        },
+        {
+            maneuverDesc: {
+                phase: "1/2",
+                ocv: "-2",
+                dcv: "+0",
+                effects: "Continuous fire on hex(es), must be Autofire",
+                attack: true,
+            },
+        },
+    );
+    addPower(undefined, {
+        key: "SWEEP",
+        type: ["maneuver"],
+        behaviors: ["non-hd", "optional-maneuver"],
+        name: "Sweep",
+        perceivability: "obvious",
+        duration: "instant",
+        costEnd: false,
+        target: "target's dcv",
+        ignoreFor: ["base2", "computer", "ai"],
+        maneuverDesc: {
+            phase: "1",
+            ocv: "-2/x",
+            dcv: "x1/2",
+            effects: "Make multiple HTH attacks",
+            attack: true,
+        },
+    });
+
+    addPower(
+        {
+            key: "THROW",
+            type: ["maneuver"],
+            behaviors: ["non-hd"],
+            name: "Throw",
+            perceivability: "obvious",
+            duration: "instant",
+            costEnd: false,
+            target: "target's dcv",
+            ignoreFor: ["base2", "computer", "ai"],
+            maneuverDesc: {
+                phase: "1/2",
+                ocv: "+0",
+                dcv: "+0",
+                effects: "Throw object or character, does STR damage",
+                attack: true,
+            },
+        },
+        undefined,
+    );
+    addPower(
+        {
+            key: "TRIP",
+            type: ["maneuver"],
+            behaviors: ["non-hd"],
+            name: "Trip",
+            perceivability: "obvious",
+            duration: "instant",
+            costEnd: false,
+            target: "target's dcv",
+            ignoreFor: ["base2", "computer", "ai"],
+            maneuverDesc: {
+                phase: "1/2",
+                ocv: "-1",
+                dcv: "-2",
+                effects: "Knock a target to the ground, making him Prone",
+                attack: true,
+            },
+        },
+        undefined,
     );
 })();
 
@@ -1283,11 +1876,10 @@ function addPower(powerDescription6e, powerOverrideFor5e) {
         {},
     );
     addPower(
-        // This is a "fake" power that doesn't exist in HD but we provide for simplicity
         {
             key: "PERCEPTION",
             type: ["skill"],
-            behaviors: ["success"],
+            behaviors: ["success", "non-hd"],
         },
         {},
     );
@@ -3142,10 +3734,9 @@ function addPower(powerDescription6e, powerOverrideFor5e) {
 
     addPower(
         {
-            // TODO: Needed?
             key: "MANEUVER",
-            type: ["martial", "attack"],
-            behaviors: ["attack"],
+            type: ["martial", "attack"], // TODO: Not all of these are attacks
+            behaviors: [],
         },
         {},
     );
@@ -3156,6 +3747,16 @@ function addPower(powerDescription6e, powerOverrideFor5e) {
             type: ["martial"],
             behaviors: [],
             costPerLevel: 4,
+        },
+        {},
+    );
+
+    addPower(
+        {
+            key: "WEAPON_ELEMENT",
+            type: ["martial"],
+            behaviors: [],
+            categorized: true,
         },
         {},
     );