+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Releases
 
-## Version 3.0.61 [Hero System 6e (Unofficial) v2](https://github.com/dmdorman/hero6e-foundryvtt)
+## Version 3.0.62 (So far...) [Hero System 6e (Unofficial) v2](https://github.com/dmdorman/hero6e-foundryvtt)
+
+- Add all 5e/6e maneuvers and optional maneuvers to actors. Will show on sheet when configuration option selected, but they may not work correctly.
+
+## Version 3.0.61
 
 - Power & Skill descriptions are now readonly as they are not intended to be edited. Descriptions are automatically generated using the Power/Skill properties. [#801](https://github.com/dmdorman/hero6e-foundryvtt/issues/801)
 - Actor shows systemVersion at time of HDC upload on actor sheet. [#800](https://github.com/dmdorman/hero6e-foundryvtt/issues/800)
@@ -9,7 +13,7 @@
 
 ## Version 3.0.60
 
-- Fixes for TELEKINESIS (TK punch/strike). 
+- Fixes for TELEKINESIS (TK punch/strike).
 - Effective STR specified as part of the attack is no longer ignored.  Allows for attacks at less than full power, thus conserving END.
 - Fixes for ENDURANCERESERVE. [#783](https://github.com/dmdorman/hero6e-foundryvtt/issues/783)
 - 5e now shows proper lift capacities for STR values under 5. [#709](https://github.com/dmdorman/hero6e-foundryvtt/issues/709)